Flask-Back/requirements.txt
numpy==1.26.4
opencv-python==4.10.0.84
pillow==10.4.0
torch==2.5.0
torchvision==0.20.0
ultralytics==8.3.57
boto3==1.34.162
python-dotenv==1.0.1
flask==3.0.0
flask-cors==4.0.0
werkzeug==3.0.1
pandas==2.2.2
waitress==3.0.0
